// types.ts
export enum GeneratorMode {
  IMAGE = 'image',
  VIDEO = 'video',
  CHARACTER = 'character',
  EDIT_IMAGE = 'edit_image',
  ANALYZE_IMAGE = 'analyze_image',
  LIVE_CHAT = 'live_chat',
  CHATBOT = 'chatbot',
  TRANSCRIBE = 'transcribe',
  PRO_TASK = 'pro_task',
  TTS = 'tts',
  NANO_RESEARCH = 'nano_research',
}